rename ResetPasswordRedesign folder to ResetPassword
