+++ packages/fxa-settings/src/components/App/index.tsx
@@ -53,12 +53,12 @@ import Legal from '../../pages/Legal';
 import LegalPrivacy from '../../pages/Legal/Privacy';
 import LegalTerms from '../../pages/Legal/Terms';
 import ThirdPartyAuthCallback from '../../pages/PostVerify/ThirdPartyAuthCallback';
-import ResetPasswordContainer from '../../pages/ResetPasswordRedesign/ResetPassword/container';
-import ConfirmResetPasswordContainer from '../../pages/ResetPasswordRedesign/ConfirmResetPassword/container';
-import CompleteResetPasswordContainer from '../../pages/ResetPasswordRedesign/CompleteResetPassword/container';
-import AccountRecoveryConfirmKeyContainer from '../../pages/ResetPasswordRedesign/AccountRecoveryConfirmKey/container';
-import ResetPasswordConfirmed from '../../pages/ResetPasswordRedesign/ResetPasswordConfirmed';
-import ResetPasswordWithRecoveryKeyVerified from '../../pages/ResetPasswordRedesign/ResetPasswordWithRecoveryKeyVerified';
+import ResetPasswordContainer from '../../pages/ResetPassword/ResetPassword/container';
+import ConfirmResetPasswordContainer from '../../pages/ResetPassword/ConfirmResetPassword/container';
+import CompleteResetPasswordContainer from '../../pages/ResetPassword/CompleteResetPassword/container';
+import AccountRecoveryConfirmKeyContainer from '../../pages/ResetPassword/AccountRecoveryConfirmKey/container';
+import ResetPasswordConfirmed from '../../pages/ResetPassword/ResetPasswordConfirmed';
+import ResetPasswordWithRecoveryKeyVerified from '../../pages/ResetPassword/ResetPasswordWithRecoveryKeyVerified';
 import CompleteSigninContainer from '../../pages/Signin/CompleteSignin/container';
 import SigninContainer from '../../pages/Signin/container';
 import ReportSigninContainer from '../../pages/Signin/ReportSignin/container';
