+++ packages/fxa-settings/src/pages/ResetPassword/CompleteResetPassword/index.test.tsx
