+++ packages/fxa-settings/src/pages/ResetPassword/CompleteResetPassword/index.tsx
