+++ packages/fxa-settings/src/pages/ResetPassword/ConfirmResetPassword/en.ftl
