+++ packages/fxa-settings/src/pages/ResetPassword/ConfirmResetPassword/index.tsx
