+++ packages/fxa-settings/src/pages/ResetPassword/ResetPassword/index.test.tsx
