+++ packages/fxa-settings/src/pages/ResetPassword/ResetPasswordConfirmed/index.stories.tsx
