+++ packages/fxa-settings/src/pages/ResetPassword/ResetPasswordConfirmed/index.test.tsx
