+++ packages/fxa-settings/src/pages/ResetPassword/ResetPasswordConfirmed/interfaces.ts
