+++ packages/fxa-settings/src/pages/ResetPassword/ResetPasswordWithRecoveryKeyVerified/mocks.tsx
